Add classification dataset adapter
Let's say we want to learn to classify images of cats.

We have lots of unlabeled images. We use these unlabeled images to
learn an encoder from image -> latent vector.
(We also learn a latent vector -> image decoder at this step as a side-effect)

Once the encoder is trained, we take a few labeled examples and learn
a simple linear classifier or MLP on the latent vector.