Train jointly with the Classifier (a new decoder)
The classifier is a decoder, but one without a discriminator.
We train the encoder and classifier jointly to ensure that the latent space
captures the important features of examples (eg. class labels)

The build_model() code is unwieldy and needs to be rewritten at this point.

Really, we should support any number of decoders, each one with or without a discriminator