Implementation of counterfactual classification
Say we have a labeled dataset of dogs and cats.

First, we learn the manifold of the images using an adversarial autoencoder.
This does not require the labels.

Then, we train a cat/dog classifier from the latent space to the labels.

Then we take some input (a dog, say) and we ask the following:
    "What would this dog look like if it were a cat instead?"

We optimize the latent vector for that input to find a latent vector that classifies as "cat".

Finally, we output the image for that new latent vector.

If all went as planned, it should be mostly the same image.
But, a small part of the image will have changed so that the dog is now replaced with a cat.