Add Attribute dataset
Combines classification with an additional binary crossentropy boolean vector

We can define boolean attributes (this person is wearing glasses, this bird
has a white breast, this number is symmetrical) along with classes.

Then during counterfactuals, we can see which attributes are related to changing
an example from one class to another.

This might allow statements like the following:
This bird is a Laysan Albatross, but if it was an Ivory Gull then it would have
    + white breast 20%
    - long neck 15%
    - long wings 10%